+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Serve outgoing public activities.
 - Added `federation.inbox_queue_batch_size` configuration parameter.
 - Check PostgreSQL version at startup.
+- Support search with `type` parameter.
 
 ### Changed